add config reload, town joining, resident list, and more

+++ src/main/kotlin/site/remlit/blueb/residential/model/Config.kt
@@ -24,7 +24,7 @@ data class DbConnectionConfig(
 data class TownConfig(
     val name: NameConfig = NameConfig(20),
     val tag: TagConfig = TagConfig(true, 4),
-    val cost: Int = 150,
+    val cost: Double = 150.0,
     val claimableChunks: ClaimableChunksConfig = ClaimableChunksConfig(500, 15, 50, 15),
     val permissions: PermissionsConfig = PermissionsConfig(1, 3, 3, 3, 2),
     val roles: RolesConfig = RolesConfig(15, listOf(
